fix: add cause message to "Caused by" lines (#100)
# Description

We use the filtered stacktrace plugin to log filtered stack traces
without a lot of Spring MVC internal classes. A coworker reported that
nested exceptions were losing their message when reported to our log
aggregator. When I looked into it, it seems that "caused by" messages
only include the exception class, not the message. This PR fixes that
issue by including the cause message in the same way that Java's native
exception printing does (capitalized "Caused by" as well as including
the message).